requirements.txt
anytree==2.13.0
chardet==5.2.0
cocotb==1.9.2
cocotb_bus==0.2.1
cocotb_coverage==1.2.0
cocotb_test==0.2.5
crc==7.1.0
matplotlib==3.10.7
numpy==2.3.4
ortools==9.12.4544
pandas==2.3.3
peakrdl_html==2.11.0
peakrdl_markdown==1.0.2
peakrdl_regblock==1.1.1
pint==0.25
psutil==5.9.8
py2puml==0.10.0
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-rerunfailures==16.1
pytest-xdist==3.6.1
rich==14.2.0
seaborn==0.13.2
simpy==4.1.1
systemrdl_compiler==1.30.1
vcdvcd==2.4.6
